# docs/openapi.yaml
openapi: 3.1.0
info:
  title: agent-identity credential resolution API
  version: 0.3.0
  description: |
    Language-agnostic REST contract for resolving agent credentials.
    Any application (Python, Go, Ruby, Java) can POST to /api/resolve or
    /api/migrate/resolve. The server resolves the credential from an encrypted
    store and returns only safe metadata — never the raw secret.

    v0.3.0 adds attestation, approval, budget, federation, and anomaly endpoints
    that were introduced in the Phase 1–4 feature build.
  contact:
    name: Datacules LLC
    url: https://github.com/hvrcharon1/agent-identity
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: http://localhost:3001
    description: Docker sidecar
paths:
...
  /api/anomaly:
    post:
      operationId: observeAnomaly
      summary: Pass an AgentRequestContext through the anomaly detector
      tags: [Anomaly]
      description: |
        Wraps a resolve() call with anomaly detection. Scores the context
        against the agent's behavioral baseline and returns any anomaly events.
        If the policy action is 'block' and a high-severity anomaly is detected,
        blocked: true is returned and the caller should not proceed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequestContext'
      responses:
        '200':
          description: Anomaly check complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyObserveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      operationId: resetAnomalyBaseline
      summary: Reset a specific agent's behavioral baseline
      tags: [Anomaly]
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: Agent user ID whose baseline to reset
      responses:
        '200':
          description: Baseline reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  schemas:
...
    AgentRequestContext:
      type: object
      required:
        - userId
        - resourceId
        - resourceKind
        - provider
        - model
        - action
        - traceId
        - requestedAt
      properties:
        userId:
          type: string
          minLength: 1
        resourceId:
          type: string
          minLength: 1
        resourceKind:
          $ref: '#/components/schemas/ResourceKind'
        provider:
          $ref: '#/components/schemas/SupportedProvider'
        model:
          type: string
        action:
          type: string
        traceId:
          type: string
        sessionId:
          type: string
        requestedAt:
          type: string
          format: date-time
        parentTraceId:
          type: string
...
    AnomalyObserveResponse:
      type: object
      required: [ok, observed, anomalies, blocked]
      properties:
        ok:
          type: boolean
        observed:
          type: string
          description: userId that was scored
        anomalies:
          type: array
          items:
            $ref: '#/components/schemas/AnomalyEvent'
        blocked:
          type: boolean
          description: True if a block-policy anomaly prevented resolution
    AnomalyEvent:
      type: object
      required: [signal, severity, baselineValue, observedValue, userId, timestamp]
      properties:
        signal:
          type: string
          enum:
            - rate_spike
            - new_credential_type
            - new_action_type
            - new_resource_kind
            - off_hours
            - new_provider
        severity:
          type: string
          enum: [low, medium, high]
        baselineValue:
          description: The known-good value or set
        observedValue:
          description: The value that triggered the anomaly
        userId:
          type: string
        credentialId:
          type: string
        timestamp:
          type: string
          format: date-time

    # ── Shared ────────────────────────────────────────────────────────────────
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ResourceKind:
      type: string
      enum: [shared, personal]
    SupportedProvider:
      type: string
      enum: [openai, anthropic, gemini, mistral, local]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Missing or invalid field in request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...